Small fixes and simplifications.

+++ bu/PointAccessor.hpp
@@ -51,7 +51,8 @@ public:
     Point operator[](size_t offset)
     {
         for (FileInfo *fi : m_fileInfos)
-            if (offset >= fi->start() && offset < fi->start() + fi->numPoints())
+            if (offset >= (size_t)fi->start() &&
+                offset < (size_t)fi->start() + fi->numPoints())
                 return Point(fi->address() + ((offset - fi->start()) * m_b.pointSize));
         return Point();
     }
